Clean up initializer and its libs
