+++ src/constants/initializer.js
@@ -1,8 +1,10 @@
 const { processMode, operationMode } = require('../environment.js');
 const { permissionLevels } = require('../constants.js');
-const { logInfo, logError } = require('../modules/errors.js');
 
 const constantsModel = require('./model.js');
+const {
+  createConstantPromise, BaseConstantsInitializer
+} = require('./initializer-lib.js');
 
 const userConstants = require('../users/constants.js');
 const userModel = require('../users/model.js');
@@ -16,84 +18,7 @@ const inventoryLib = require('../inventory/lib.js');
 const inventoryConstants = require('../inventory/constants.js');
 
 
-/*
- * createConstantPromise
- *  constantName [String]: unique name of the constant to create
- *  creationFn [Fn]:
- *    function that takes constant's data, returns created mongoose document
- *  data [Object]: data to create the constant with
- */
-const createConstantPromise = (
-  constantName, type, creationFn, data,
-) => {
-  return new Promise((resolve, reject) => {
-    creationFn(data)
-      .then((res) => {
-        if(! res) {
-          reject(new Error('Failed to create constant:', constantName, data, res));
-        }
-        if(res._id) {
-          constantsModel.createConstant({
-            name: constantName,
-            id: res._id,
-            type: type,
-          }).then(() => {
-            logInfo('[+] Created constant:', constantName);
-            resolve([constantName, res]);
-          })
-          .catch(err => {
-            logError('[!] Error creating Constant mongoose document');
-            reject(err);
-          });
-        } else {
-          reject(new Error('Failed to create constant:', constantName, data, res.error));
-        }
-      })
-      .catch(err => {
-        logError('[!] Error creating', constantName, ':', err.message);
-        reject(err);
-      });
-  });
-};
-
-const adminUserConstantInitializer = () => createConstantPromise(
-  'adminUser', 'user',
-  (userData) => userLib.createUser(userData),
-  userConstants.adminUserData,
-)
-
-class baseConstantsInitializer {
-  constructor(){
-    /*
-     * initializers [Map, constantName [String] : initializerFn [null => Promise]
-     */
-    this.initializers = {
-      'adminUser': adminUserConstantInitializer,
-    };
-    /*
-     * postProcessors [Array, postProcessFn]
-     *               - Function, (resultMap [Map, contantName: document]) => Promise
-     */
-    this.postProcessors = [
-      // set adminUser's permissions to Admin
-      (resultMap) => new Promise((resolve, reject) => {
-        if(Object.hasOwnProperty.call(resultMap, 'adminUser')){
-          userModel.patchUser(
-            resultMap['adminUser']._id,
-            {permissionLevel: permissionLevels.ADMIN},
-          )
-            .then(resolve)
-            .catch(reject)
-        } else {
-          resolve()
-        }
-      }),
-    ];
-    this.postSetup = [];
-  }
-}
-
-class ambassadorsiteConstantsInitializer extends baseConstantsInitializer {
+class AmbassadorsiteConstantsInitializer extends BaseConstantsInitializer {
   constructor() {
     super();
     this.initializers['ambassadorApplication'] = () => createConstantPromise(
@@ -124,7 +49,8 @@ class ambassadorsiteConstantsInitializer extends baseConstantsInitializer {
   }
 }
 
-class stocktrackerConstantsInitializer extends baseConstantsInitializer {
+
+class StocktrackerConstantsInitializer extends BaseConstantsInitializer {
   constructor() {
     super();
     this.postSetup = inventoryLib.postSetup;
@@ -275,6 +201,6 @@ class stocktrackerConstantsInitializer extends baseConstantsInitializer {
 }
 
 module.exports = {
-  ambassadorsite: ambassadorsiteConstantsInitializer,
-  stocktracker: stocktrackerConstantsInitializer,
+  ambassadorsite: AmbassadorsiteConstantsInitializer,
+  stocktracker: StocktrackerConstantsInitializer,
 }[processMode];
